requirements.txt
openai>=1.0.0
streamlit>=1.20.0
python-dotenv>=0.19.0
sentence-transformers>=2.0.0
scikit-learn
numpy
pandas
